Refactor: Split build.py into smaller protocol-based modules
- Created `build_protocols` directory to house new modules.
- Moved distinct functionalities (translation, data loading, HTML generation, config management, page assembly) into separate Python files within `build_protocols`.
- Updated `build.py` to delegate logic to these new modules, simplifying its role to orchestration.
- Cleaned up imports in `build.py`.
- Added `sys.path` adjustments to `test_build.py` to aid in locating generated protobuf files.

Note: Testing was hindered by an issue with populating the `generated/` directory via `npm run generate-proto` in the test environment. The Python refactoring itself is complete.
