+++ build_protocols/__init__.py
@@ -1,0 +1,10 @@
+# This file makes Python treat the directory as a package.
+# It can also be used to define what symbols are exported when
+# 'from build_protocols import *' is used, though that's not strictly necessary here.
+
+# You can also make sub-modules available more directly, e.g.:
+# from .translation import load_translations
+# from .data_loading import load_dynamic_data
+# etc.
+# For now, we'll keep it simple and rely on direct imports in build.py
+# e.g., from build_protocols.translation import load_translations
